+++ fastf1/legacy.py
@@ -25,7 +25,7 @@ versions can be used.
     fastf1.Cache.enable_cache("test_cache/")
 
     session = fastf1.get_session(2020, 'Italy', 'R')
-    laps = session.load_laps()
+    laps = session.load_laps(with_telemetry=True)
 
     DRIVER = 'VER'  # which driver; need to specify number and abbreviation
     DRIVER_NUMBER = '33'