+++ fastf1/livetiming/__init__.py
@@ -42,7 +42,7 @@ Usage Example
 
     livedata = LiveTimingData('saved_data.txt')
     session = fastf1.get_session(2021, 'testing', 1)
-    session.load_laps(livedata=livedata)
+    session.load_laps(with_telemetry=True, livedata=livedata)
 
 optionally you can load live timing data from two or more files